Do not modify /etc/localtime on every boot
Only create or change the symlink when TIMEZONE has a value for the
first time, or when its value doesn't match the current /etc/localtime
symlink. This should prevent an error when /etc/localtime is on a read-only
filesystem, and should prevent a false alarm from an audit tool noticing
a change in its timestamp. It should also help reduce wear on a
filesystem stored on Flash memory.